LayoutTests/fast/canvas/canvas-context-save-limit-expected.txt

CONSOLE MESSAGE: line 29: CanvasRenderingContext2D.save() has been called without a matching restore() too many times. Ignoring save().
CONSOLE MESSAGE: line 29: CanvasRenderingContext2D.save() has been called without a matching restore() too many times. Ignoring save().
This test verifies that the size of CanvasRenderingContext2D state stack does not grow beyond a certain limit.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS ctx.fillStyle is "#ff0000"
PASS ctx.fillStyle is "#0000ff"
PASS ctx.fillStyle is "#00ff00"
PASS successfullyParsed is true

TEST COMPLETE
